Move NPCycle and CycleLog
